Prepare a 1.0a4 release.
The primary bug was caused by always setting LoginInfo.formSubmitURL to the same value as LoginInfo.hostname. This blew up for forms that specified a different action URL field because the password manager actually does a comparison even after searching. Instead we always set it to "" (empty string) which currently matches any value in LoginInfo.matches().

Fixing this also brought up some confusion around the meaning of null and "" as passed to the findLogin() interface. "" should mean match any value, and null means match no value. Adjust _findKeychainItems() to expect the same pattern and to convert the values as appropriate before calling the keychain API calls.

Also standardize on using the presence of a non-null value for LoginInfo.formSubmitURL to indicate that this is a "form" login. All other logins are "non-form" logins (currently stored as the default type).

Adjusted a few comments and other lines to 80 chars.