Update compatibility to indicate that the extension works with Thunderbird 9 as reported by several users.
Versions of TB and FF >= 10.0 will automatically install extensions even if the maxVersion in install.rdf is below the installed version of the application, so this should finally allow TB users to be able to install the extension on a regular basis.

Note that TB still will not share credentials with Apple's Mail.app. This is because Mail.app includes the port number in the keychain item and TB's imap URL parser fails to parse imap:// URLs that contain port numbers. TB also generates imap:// urls even when IMAPS is being used.